module-08/docker/flask-llm-app/requirements.txt
flask
langchain
langchain-openai
python-dotenv
azure-identity
langchain_classic
gunicorn